backend/requirements.txt
Django==6.0.4
django-cors-headers==4.9.0
django-webpush==0.3.6
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
google-auth==2.49.1
google-auth-oauthlib==1.3.1
gunicorn==23.0.0
whitenoise==6.8.2
dj-database-url==2.3.0
psycopg2-binary==2.9.10
